requirements.txt
fastapi
uvicorn
databases
sqlalchemy
aioredis
python-binance
pyjwt
python-dotenv
alembic
psycopg2-binary
email-validator
fastapi-mail

